+++ src/xnnpack/simd/s32-avx512f.h
@@ -61,7 +61,7 @@ static XNN_INLINE void xnn_storeu_s32(int32_t* ptr, xnn_simd_s32_t v) {
   _mm512_storeu_epi32(ptr, v);
 }
 
-static XNN_INLINE void xnn_store_s32(float* ptr, xnn_simd_s32_t v) {
+static XNN_INLINE void xnn_store_s32(int32_t* ptr, xnn_simd_s32_t v) {
   _mm512_store_epi32(ptr, v);
 }
